Rename @ndla/dropdown to @ndla/dropdown-menu
