+++ packages/dropdown-menu/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@ndla/dropdown",
+  "name": "@ndla/dropdown-menu",
   "version": "0.1.0",
-  "description": "Dropdown component for NDLA",
+  "description": "Dropdown menu component for NDLA",
   "license": "GPL-3.0",
   "main": "lib/index.js",
   "module": "es/index.js",
@@ -15,7 +15,7 @@
   "repository": {
     "type": "git",
     "url": "https://github.com/NDLANO/frontend-packages.git",
-    "directory": "packages/dropdown"
+    "directory": "packages/dropdown-menu"
   },
   "keywords": [
     "ndla"
