+++ packages/dropdown-menu/src/index.ts
@@ -1,4 +1,4 @@
-export { default as Dropdown } from './Dropdown';
+export { default as DropdownMenu } from './DropdownMenu';
 export { default as DropdownContent } from './DropdownContent';
 export { default as DropdownItem } from './DropdownItem';
 export { default as DropdownTrigger } from './DropdownTrigger';
