--- packages/dropdown/README.md
@@ -1,25 +1,0 @@
-# ndla-dropdown
-
-Dropdown component based on @radix-ui/react-dropdown-menu
-
-## Installation
-
-```sh
-yarn add --save @ndla/dropdown
-```
-
-```sh
-npm install @ndla/dropdown --save
-```
-
-## Usage
-
-Only DropdownContent is styled, and can be targeted as one usually does with emotion. The content arrow indicator can be styled by targeting `[data-arrow]`.
-
-### Basic example
-
-```js
-import Button from '@ndla/button';
-
-<Button onClick={() => alert('Hello world')}>Click me! </Button>;
-```
